Refactor utils into organizable services
