+++ services/auth/auth.js
@@ -2,12 +2,12 @@
 // Description: Provides functions that handle the authentication process.
 
 const assert = require("assert");
-const needle = require("needle");
 const jwt = require("jsonwebtoken");
+const needle = require("needle");
 const url = require("url");
 const xml2js = require("xml2js");
 
-const {Role} = require("../entities/role");
+const {Role} = require("../../entities/role");
 
 const JWT_SECRET_KEY = process.env.JWT_SECRET_KEY;
