+++ services/validation/planValidation.js
@@ -1,7 +1,7 @@
 // File: planValidation.js
 // Description: validates a submitted plan against a list of constraints
 
-const pool = require("./mysqlPool").pool;
+const pool = require("../db/mysqlPool").pool;
 
 const CREDITS_MIN = 32;
